backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
openpyxl
slowapi